Initial upload of function volcano growth model
The volcanic edifice grows from a set of random eruption sites distributed within a certain radius of the central point. Whether eruption is possible or not depends on the state of stress (see McGovern et al., 2013). If possible, an eruption site forms a small cone of volcanic products that further flexes the lithosphere, modifying the state of stress. If not, the last eruption site is reused.
The main script is VG.m which calls all necessary attendent scripts. See VG.m for a list and description of variable.
Note this is a very preliminary commit and we expect the code to be significantly modified. At this point, it comes with no guarantee.